backend/requirements.txt
# Web Framework & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database & Environment
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# Security & Auth
cryptography==42.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# AI & Analysis (GPT-4o-mini 연동 핵심)
openai

# Data Collection & Parsing
requests==2.31.0
httpx==0.26.0
beautifulsoup4==4.12.3
yfinance>=0.2.50
pykrx>=1.0.45

# Data Science & ML (주가 예측용)
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0

# Task Scheduling
apscheduler==3.10.4